[coq] Adapt to coq/coq#9129 "removal of imperative proof state"
A note on the porting of the plugin, paramcoq is intrinsically
designed to use the global environment as it is where it adds the
definitions on the fly.

All previous uses of `get_current_context` where bogus and only worked
because there was no open proof when the commands where used.