Adapt to Coq PR#13143: Pp.h does not take a (dummy) integer anymore.
The h-boxes were called here with a very high integer. It might have
been by irony knowing that the argument is dummy or to move the
printing very far in the margin because it is intended to be debug
printing to put far in the margin. The values were 100 and 140, so I
assumed it was intended and I replaced the h-box by an hov-box.